% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{abbrev}
\alias{abbrev}
\title{Give the whole abbreviation definition}
\usage{
abbrev(param)
}
\arguments{
\item{param}{abbreviation term}
}
\value{
characater abbreviation defintion
}
\description{
Give the whole abbreviation definition
}
\examples{
abbrev("VPC")
}
