% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_draft}
\alias{add_draft}
\title{Add draft layer annotation}
\usage{
add_draft(label = "DRAFT", size, color)
}
\arguments{
\item{label}{draft layer default to DRAFT}

\item{size}{size of the annotation}

\item{color}{color of the annotation default to grey50}
}
\value{
ggplot2 annotation
}
\description{
This is add the word draft to certain graphics.
}
\examples{
add_draft("DRAFT", size = 5, color = "grey50")
}
