% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{check_shrink}
\alias{check_shrink}
\title{Performs checks of names in shrink list}
\usage{
check_shrink(shrink_list)
}
\arguments{
\item{shrink_list}{\code{list} list of shrink arguments}
}
\description{
Performs checks of names in shrink list
}
