% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tables.R
\name{combine_tables}
\alias{combine_tables}
\title{Combine tables}
\usage{
combine_tables(x)
}
\arguments{
\item{x}{A list containing the tables (`x$data`) to be 
combined, their names (`x$name`), their number of rows (`x$nrow`) 
and their respective index (`x$index`).}
}
\value{
A list containing `data` and `index` of the combined table.
}
\description{
Combine tables
}
\keyword{internal}
