% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configs}
\alias{configs}
\title{List all vaiable config}
\usage{
configs(sys = "mlx")
}
\arguments{
\item{sys}{can be mlx, if missed all configurations will be listed}
}
\value{
names of the config
}
\description{
List all vaiable config
}
\examples{
configs()
}
