% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distribution.R
\name{distrib}
\alias{distrib}
\title{Create Graphic distribution object}
\usage{
distrib(labels, has.jitter = TRUE, jitter = list(shape = 2, color =
  "grey50", width = 1), facets = list(scales = "free", nrow = 3),
  type = "box", has.shrink = FALSE, ...)
}
\arguments{
\item{labels}{list of texts/titles used within the plot}

\item{jitter}{list set jitter parameter}

\item{facets}{list set the facet setting in case of histogram plot}

\item{type}{box for boxplot or histogram}

\item{...}{others graphics arguments passed to \code{\link{pmx_gpar}} internal object.}
}
\value{
gdistribution object
}
\description{
Create Graphic distribution object
}
\seealso{
\code{\link{plot_pmx.distrib}}
}
