% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforce-facet-paginate.R
\docType{data}
\name{FacetWrapPaginate}
\alias{FacetWrapPaginate}
\title{Extend facet_wrap using ggproto}
\usage{
FacetWrapPaginate
}
\description{
Extend facet_wrap using ggproto
}
\keyword{datasets}
