% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem_reader.R
\name{file_path}
\alias{file_path}
\title{Generate clean file paths}
\usage{
file_path(dir, file)
}
\arguments{
\item{dir}{A string or vector of strings containing the directory path.}

\item{file}{A file name or vector containing the file names.}
}
\value{
A string or vector of string of the full file path.
}
\description{
Wrapper around `file.path` that cleans trailing forward 
slash and missing `dir`.
}
\keyword{internal}
