% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_abbrev}
\alias{get_abbrev}
\title{Get abbreviation definition by key}
\usage{
get_abbrev(ctr, param)
}
\arguments{
\item{ctr}{\code{pmxClass} controller}

\item{param}{abbreviation term}
}
\value{
\code{character} abbreviation definition
}
\description{
Get abbreviation definition by key
}
