% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_conts}
\alias{get_conts}
\title{Get continuous covariates}
\usage{
get_conts(ctr)
}
\arguments{
\item{ctr}{the controller object}
}
\value{
a charcater vector
}
\description{
Get continuous covariates
}
\seealso{
Other pmxclass: 
\code{\link{get_cats}()},
\code{\link{get_covariates}()},
\code{\link{get_data}()},
\code{\link{get_occ}()},
\code{\link{get_plot}()},
\code{\link{get_plot_config}()},
\code{\link{get_strats}()},
\code{\link{plot_names}()},
\code{\link{plots}()},
\code{\link{pmx_update}()},
\code{\link{set_data}()},
\code{\link{set_plot}()}
}
\concept{pmxclass}
