% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-all-plots.R
\name{get_params_from_call}
\alias{get_params_from_call}
\title{Get list of parameters based on parameters used in own function call}
\usage{
get_params_from_call()
}
\description{
Get list of parameters based on parameters used in own function call
}
