% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_plot}
\alias{get_plot}
\title{Get plot object}
\usage{
get_plot(ctr, nplot, npage)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}

\item{nplot}{characater the plot name}

\item{npage}{integer or integer vector, set page number in case of multi pages plot}
}
\value{
ggplot object
}
\description{
Get plot object
}
