% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_plot_config}
\alias{get_plot_config}
\title{Get the plot config name}
\usage{
get_plot_config(ctr, pname)
}
\arguments{
\item{ctr}{the controller object}

\item{pname}{the plot name}
}
\value{
the config object
}
\description{
Get the plot config name
}
\examples{
\dontrun{
pmxOptions(work_dir=WORK_DIR)
ctr <- pmx_mlx(config = "standing")
ctr \%>\% set_plot("IND", pname = "indiv1")
get_plot_config("distr1")
}
}
