% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_strats}
\alias{get_strats}
\title{Get extra stratification variables}
\usage{
get_strats(ctr)
}
\arguments{
\item{ctr}{the controller object}
}
\value{
a charcater vector
}
\description{
Get extra stratification variables
}
\seealso{
Other pmxclass: \code{\link{get_cats}},
  \code{\link{get_conts}}, \code{\link{get_covariates}},
  \code{\link{get_data}}, \code{\link{get_occ}},
  \code{\link{get_plot_config}}, \code{\link{get_plot}},
  \code{\link{plot_names}}, \code{\link{plots}},
  \code{\link{pmx_update}}, \code{\link{set_data}},
  \code{\link{set_plot}}
}
\concept{pmxclass}
