% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPMX-options.R
\name{getPmxOption}
\alias{getPmxOption}
\title{Get ggPMX Option}
\usage{
getPmxOption(name, default = NULL)
}
\arguments{
\item{name}{Name of an option to get.}

\item{default}{Value to be returned if the option is not currently set.}
}
\value{
value of the option or NULL
}
\description{
Get ggPMX Option
}
\examples{
\dontrun{
pmxOptions(myOption = 10)
getPmxOption("myOption")
}
}
