% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpmx_quoted.R
\name{ggpmx_quoted}
\alias{ggpmx_quoted}
\title{Convert input to quoted variables.}
\usage{
ggpmx_quoted(x, env = parent.frame())
}
\arguments{
\item{x}{input to quote}

\item{env}{environment in which unbound symbols in expression should be
evaluated. Defaults to the environment in which \code{ggpmx_quoted} was
executed.}
}
\value{
a list of quoted variables
}
\description{
Convert characters, formulas and calls to quoted .variables
}
\details{
This method was taken from `plyr` so it wouldn't be an additional dependency

Currently conversions exist for character vectors, formulas and
call objects.
}
\examples{
ggpmx_quoted(c("a", "b", "log(d)"))
ggpmx_quoted(a ~ b + log(d))
}
\seealso{
\code{\link[=quoted]{.}}
}
