% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual.R
\name{individual}
\alias{individual}
\title{Create individual fits object}
\usage{
individual(labels, has.curve = TRUE, facets = list(ncol = 3, nrow = 4), ...)
}
\arguments{
\item{labels}{plot tesxts. labels, axis,}

\item{has.curve}{logical if TRUE add predictions lines}

\item{facets}{list facets settings nrow/ncol}

\item{...}{others graphics arguments passed to \code{\link{pmx_gpar}} internal object.}
}
\value{
individual fit object
}
\description{
Create individual fits object
}
\seealso{
\code{\link{plot_pmx.individual}}
}
