% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-post_load.R
\name{input_finegrid}
\alias{input_finegrid}
\title{Merge input and fingrid data sets}
\usage{
input_finegrid(input, finegrid)
}
\arguments{
\item{input}{\code{data.table} input data set}

\item{finegrid}{\code{data.table} finegrid data set}
}
\value{
data.table
}
\description{
Merge input and fingrid data sets
}
