% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{is.pmx_gpar}
\alias{is.pmx_gpar}
\title{Check if an object is a pmx_gpar class}
\usage{
is.pmx_gpar(x)
}
\arguments{
\item{x}{pmx_gpar object}
}
\value{
\code{logical} returns TRUE if it is a \code{pmx_gpar} object
}
\description{
Check if an object is a pmx_gpar class
}
