% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_character_or_formula_or_null_or_na}
\alias{is_character_or_formula_or_null_or_na}
\title{In contrast to is_string_or_formula_or_null returns TRUE for c('A', 'B')
Also c('A', NA) ~ TRUE, NA ~ TRUE}
\usage{
is_character_or_formula_or_null_or_na(x)
}
\description{
In contrast to is_string_or_formula_or_null returns TRUE for c('A', 'B')
Also c('A', NA) ~ TRUE, NA ~ TRUE
}
