% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_left_join}
\alias{l_left_join}
\title{Merge 2 lists}
\usage{
l_left_join(base_list, overlay_list, recursive = TRUE)
}
\arguments{
\item{base_list}{list to update}

\item{overlay_list}{list used to update the first list}

\item{recursive}{logical if TRUE do the merge in depth}
}
\value{
list
}
\description{
left join , the first list is updated by the seond one
}
