% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nm_tables.R
\name{list_nm_tables}
\alias{list_nm_tables}
\title{List NONMEM output tables}
\usage{
list_nm_tables(nm_model = NULL)
}
\arguments{
\item{nm_model}{An xpose nm_model object generated with \code{\link{read_nm_model}}.}
}
\description{
List NONMEM output tables file names from a \code{nm_model} object.
}
\examples{
\dontrun{
read_nm_model(file = 'run001.lst') \%>\% 
  list_nm_tables()
}

}
\seealso{
\code{\link{read_nm_model}}, \code{\link{read_nm_tables}}
}
