% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_import.R
\name{list_nm_tables_manual}
\alias{list_nm_tables_manual}
\title{Creates an nm_table_list from manually defined table name patterns}
\usage{
list_nm_tables_manual(runno = NULL, file = NULL, dir = NULL,
  tab_list)
}
\arguments{
\item{runno}{Run number to be used to generate model file name.}

\item{file}{Model file name containing the file extension.}

\item{dir}{Location of the model files.}

\item{tab_list}{A list of table definition generated by `manual_nm_import`.}
}
\value{
A `nm_table_list`
}
\description{
Creates an nm_table_list from manually defined table name patterns
}
\keyword{internal}
