% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Get data source config}
\usage{
load_config(x, sys)
}
\arguments{
\item{x}{the config name.}

\item{sys}{can be mlx,nm,...}
}
\value{
a list :data configuration object
}
\description{
Get data source config
}
