% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{load_source}
\alias{load_source}
\title{Load all/or some source data set}
\usage{
load_source(sys, path, dconf, include, exclude)
}
\arguments{
\item{sys}{type cane mlx/nom}

\item{path}{\code{character} directory path containing all sources.}

\item{dconf}{configuration object}

\item{include}{list of data sets to load,
if missed ( the default behavior) load all sources.}
}
\value{
list of data.table
}
\description{
Load all/or some source data set
}
