% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem_reader.R
\name{make_extension}
\alias{make_extension}
\title{Generate extension string}
\usage{
make_extension(x)
}
\arguments{
\item{x}{A string or vector of strings containing the extension to be standardized.}
}
\value{
A string or vector of strings of extension(s).
}
\description{
Generate consistent extension strings by adding dot 
prefix whenever necessary.
}
\keyword{internal}
