% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforce-facet-paginate.R
\name{n_pages}
\alias{n_pages}
\title{Determine the number of pages in a paginated facet plot}
\usage{
n_pages(plot)
}
\arguments{
\item{plot}{A ggplot object using either facet_wrap_paginate or
facet_grid_paginate}
}
\value{
If the plot uses  using either facet_wrap_paginate or
facet_grid_paginate it returns the total number of pages. Otherwise it
returns NULL
}
\description{
This is a simple helper that returns the number of pages it takes to plot all
panels when using \code{\link{facet_wrap_paginate}} . It partially builds the plot so depending
on the complexity of your plot it might take some time to calculate...
}
