% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_table.R
\name{param_table}
\alias{param_table}
\title{Creates parameter table}
\usage{
param_table(ctr, fun, return_table = FALSE)
}
\arguments{
\item{ctr}{The controller generated by pmx/pmx_mlx/pmx_mlxtran/pmx_nlmixr/pmx_nm}

\item{fun}{\code{character} can be "sd" or "var"}

\item{print_tibble}{If TRUE, returns tibble if FALSE returns kable}
}
\value{
Returns an improved tibble or a kable with the parameter estimates from ctr %>% get_data('estimates')
}
\description{
Creates parameter table
}
\examples{
#ctr <- theophylline()
#my_params <- ctr \%>\% param_table(fun = "sd")
}
