% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_mlxtran}
\alias{parse_mlxtran}
\title{Parse MONOLIX mlxtran file}
\usage{
parse_mlxtran(file_name)
}
\arguments{
\item{file_name}{absolute path to mlxtran file}
}
\value{
\code{list} key/values to initialize ggPMX controller
}
\description{
Parse MONOLIX mlxtran file
}
