% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pk_occ}
\alias{pk_occ}
\title{Creates pmx controller using monlix data having Occasional variable}
\usage{
pk_occ()
}
\value{
pmx controller
}
\description{
Creates pmx controller using monlix data having Occasional variable
}
\examples{
\dontrun{
pk_occ()
}
}
