% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pk_pd}
\alias{pk_pd}
\title{Creates pkpd pmx controller using package internal data}
\usage{
pk_pd(code = "3")
}
\arguments{
\item{code}{can be 3 or 4}
}
\description{
Creates pkpd pmx controller using package internal data
}
