% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distribution.R
\name{plot_pmx.distrib}
\alias{plot_pmx.distrib}
\title{Plot EBE distribution}
\usage{
\method{plot_pmx}{distrib}(x, dx)
}
\arguments{
\item{x}{distribution object}

\item{dx}{data set}
}
\value{
ggplot2 plot
}
\description{
Plot EBE distribution
}
\seealso{
\code{\link{distrib}}

Other plot_pmx: \code{\link{plot_pmx.individual}},
  \code{\link{plot_pmx.residual}}, \code{\link{plot_pmx}}
}
