% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eta-cov.R
\name{plot_pmx.eta_cov}
\alias{plot_pmx.eta_cov}
\title{This plots an ETA covariance matrix which can be used to define the co-relation between the parameters and
its shrinkage}
\usage{
\method{plot_pmx}{eta_cov}(x, dx, ...)
}
\arguments{
\item{x}{eta_cov object}

\item{dx}{data set}

\item{...}{not used for the moment}
}
\value{
ggplot2 plot
}
\description{
This plots an ETA covariance matrix which can be used to define the co-relation between the parameters and
its shrinkage
}
\seealso{
\code{\link{eta_cov}}

Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()},
\code{\link{plot_pmx}()}
}
\concept{plot_pmx}
