% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eta-pairs.R
\name{plot_pmx.eta_pairs}
\alias{plot_pmx.eta_pairs}
\title{Plot random effect correlation plot}
\usage{
\method{plot_pmx}{eta_pairs}(x, dx, ...)
}
\arguments{
\item{x}{distribution object}

\item{dx}{data set}

\item{...}{not used for the moment}
}
\value{
ggpairs plot
}
\description{
Plot random effect correlation plot
}
\seealso{
\code{\link{distrib}}

Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_param_history}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()}
}
\concept{plot_pmx}
