% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual.R
\name{plot_pmx.individual}
\alias{plot_pmx.individual}
\title{Plot individual fits}
\usage{
\method{plot_pmx}{individual}(x, dx, include)
}
\arguments{
\item{x}{individual object}

\item{dx}{data set}

\item{include}{list of individual ID to plot, if missing all ID will be plotted}
}
\value{
a list of ggplot2
}
\description{
Plot individual fits
}
\seealso{
\code{\link{individual}}

Other plot_pmx: \code{\link{plot_pmx.distrib}},
  \code{\link{plot_pmx.residual}}, \code{\link{plot_pmx}}
}
