% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density.R
\name{plot_pmx.pmx_dens}
\alias{plot_pmx.pmx_dens}
\title{This function plots EBE versus covariates using qq plots}
\usage{
\method{plot_pmx}{pmx_dens}(x, dx, ...)
}
\arguments{
\item{x}{eta_cov object}

\item{dx}{data set}

\item{...}{not used for the moment}
}
\value{
ggplot2 plot
}
\description{
This function plots EBE versus covariates using qq plots
}
\seealso{
\code{\link{eta_cov}}

Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()},
\code{\link{plot_pmx}()}
}
\concept{plot_pmx}
