% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-base.R
\name{plot_pmx.pmx_gpar}
\alias{plot_pmx.pmx_gpar}
\title{The ggPMX base plot function}
\usage{
\method{plot_pmx}{pmx_gpar}(gpar, p)
}
\arguments{
\item{gpar}{object of pmx_gpar type}

\item{p}{plot}
}
\value{
ggplot2 object
}
\description{
This function should be called internally by other plots to set
general settings like , smoothing, add band, labelling, theming,...
}
