% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-saem-convergence.R
\name{plot_pmx.pmx_param_history}
\alias{plot_pmx.pmx_param_history}
\title{S3 method for plots of class pmx_param_history}
\usage{
\method{plot_pmx}{pmx_param_history}(x, dx, ...)
}
\arguments{
\item{x}{pmx_param_history object}

\item{dx}{data set}

\item{...}{not used}
}
\value{
ggplot2 plot
}
\description{
S3 method for plots of class pmx_param_history
}
\seealso{
Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()}
}
\concept{plot_pmx}
