% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pmx-plots-vpc.R
\name{plot_pmx}
\alias{plot_pmx}
\alias{plot_pmx.pmx_vpc}
\title{This is a generic plot method that produces all plots by default described in pmx model
evaluation guidance.}
\usage{
plot_pmx(x, dx, ...)

\method{plot_pmx}{pmx_vpc}(x, dx, ...)
}
\arguments{
\item{x}{object to plot}

\item{dx}{data.table , plot source data}

\item{...}{extra argument (not used)}
}
\description{
This is a generic plot method that produces all plots by default described in pmx model
evaluation guidance.
}
\seealso{
\code{\link{pmx_gpar}}.

Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_param_history}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()}
}
\concept{plot_pmx}
