% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residual.R
\name{plot_pmx.residual}
\alias{plot_pmx.residual}
\title{Plot residual object}
\usage{
\method{plot_pmx}{residual}(x, dx)
}
\arguments{
\item{x}{residual object}

\item{dx}{data set}
}
\value{
ggplot2 object
}
\description{
Plot residual object
}
\seealso{
\code{\link{residual}}

Other plot_pmx: \code{\link{plot_pmx.distrib}},
  \code{\link{plot_pmx.individual}}, \code{\link{plot_pmx}}
}
