% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-eta-pairs.R
\name{plot_shrink}
\alias{plot_shrink}
\title{Plot shrink in eta matric}
\usage{
plot_shrink(x, shrink.dx, shrink)
}
\arguments{
\item{x}{pmx_gpar object}

\item{shrink.dx}{data.table of shrinkage}

\item{shrink}{\code{pmxShrinkClass} shrinkage graphical parameter or
\code{list} coercible into one}
}
\value{
ggplot2 object
}
\description{
Plot shrink in eta matric
}
