% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{plots}
\alias{plots}
\title{Get plots description}
\usage{
plots(ctr)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}
}
\value{
data.frame of plots
}
\description{
Get plots description
}
\seealso{
Other pmxclass: 
\code{\link{get_cats}()},
\code{\link{get_conts}()},
\code{\link{get_covariates}()},
\code{\link{get_data}()},
\code{\link{get_occ}()},
\code{\link{get_plot_config}()},
\code{\link{get_plot}()},
\code{\link{get_strats}()},
\code{\link{plot_names}()},
\code{\link{pmx_update}()},
\code{\link{set_data}()},
\code{\link{set_plot}()}
}
\concept{pmxclass}
