% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotTester}
\alias{plotTester}
\title{Shiny application to test ggPMX plots}
\usage{
plotTester()
}
\value{
This function does not return. It is a wrapper for runApp. 
Interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\description{
Shiny application to test ggPMX plots
}
