% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{pmx}
\alias{pmx}
\title{Create a pmx object data Source}
\usage{
pmx(config, sys, directory)
}
\arguments{
\item{config}{Can be either :
The complete path for the configuration file, the name of configuration within the built-in
list of configurations, or a configuration object.}

\item{sys}{the system name can be MLX/NM/OTHERS}

\item{directory}{where the files are located. This is an optional parameter by default pmw will look
in \code{work_dir} pmx options: \code{getPmxOptions("work_dir")}}
}
\value{
a pmxClass object
}
\description{
Create a pmx object data Source
}
\examples{

}
