% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{pmx_bloq}
\alias{pmx_bloq}
\title{Creates BLOQ object attributes}
\usage{
pmx_bloq(
  cens = "CENS",
  limit = "LIMIT",
  colour = "pink",
  size = 2,
  linewidth = 1,
  alpha = 0.9,
  show = TRUE,
  ...
)
}
\arguments{
\item{cens}{\code{character} the censoring column name}

\item{limit}{\code{character}  the limit column name (optional)}

\item{colour}{\code{character}  the color of the geom}

\item{size}{\code{numeric}  the size of the geom when using \code{geom_point()}}

\item{linewidth}{\code{numeric} the line width of the segment when using \code{geom_segment()}}

\item{alpha}{\code{numeric}  the alpha of the geom}

\item{show}{\code{logical} if FALSE remove all censory observations}

\item{...}{any other graphical parameter}
}
\description{
Creates BLOQ object attributes
}
\details{
To define that a measurement is censored, the observation data set should include
a CENSORING column ( default to `CENS` ) and put 1 for lower limit or -1 for upper limit. \cr
Optionally, data set can contain have a limit column ( default to `LIMIT`) column to set the other limit.
}
