% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_tables.R
\name{pmx_combine_tables}
\alias{pmx_combine_tables}
\title{Combine tables}
\usage{
pmx_combine_tables(x)
}
\arguments{
\item{x}{A list containing the tables (`x$data`) to be 
combined, their names (`x$name`), their number of rows (`x$nrow`) 
and their respective index (`x$index`).}
}
\value{
A list containing `data` and `index` of the combined table.
}
\description{
Combine tables
}
\keyword{internal}
