% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-all-compute.R
\name{pmx_comp_shrink}
\alias{pmx_comp_shrink}
\title{Compute Shrinkage}
\usage{
pmx_comp_shrink(ctr, fun = c("sd", "var"), strat.facet, strat.color,
  filter, ...)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}

\item{fun}{\code{character} can be sd or var}

\item{strat.facet}{\code{formula} optional stratification parameter}

\item{strat.color}{\code{character} optional stratification parameter}

\item{filter}{optional filter which will be applied to plotting data}

\item{...}{others parameters not used for the moment}
}
\value{
\code{data.table}
}
\description{
Compute Shrinkage
}
