% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_file_path}
\alias{pmx_file_path}
\title{Generate clean file paths}
\usage{
pmx_file_path(dir, file)
}
\arguments{
\item{dir}{A string or vector of strings containing the directory path.}

\item{file}{A file name or vector containing the file names.}
}
\value{
A string or vector of string of the full file path.
}
\description{
Wrapper around `file.path` that cleans trailing forward 
slash and missing `dir`.
}
\keyword{internal}
