% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pmx_get_configs}
\alias{pmx_get_configs}
\title{Get List of built-in configurations}
\usage{
pmx_get_configs(sys = "mlx")
}
\arguments{
\item{sys}{can be mlx, by default all configurations will be listed}
}
\value{
names of the config
}
\description{
Get List of built-in configurations
}
\examples{
pmx_get_configs()
}
