% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_get_extension}
\alias{pmx_get_extension}
\title{Get file extension}
\usage{
pmx_get_extension(x, dot = TRUE)
}
\arguments{
\item{x}{A string or vector of strings containing the filenames with the extension.}

\item{dot}{Logical, if `TRUE` the returned value will contain the 
dot (e.g `.mod`) else only the extension itself will be returned (e.g. `mod`).}
}
\value{
A string or vector of string of the file(s) extension.
}
\description{
Extract file extension from the filename string.
}
\keyword{internal}
