% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{pmx_gpar}
\alias{pmx_gpar}
\title{Handling pmx Graphical parameters}
\usage{
pmx_gpar(labels, axis.title = c(size = 12), axis.text = c(size = 14),
  ranges = NULL, has.smooth = TRUE, smooth = list(se = FALSE, linetype =
  2, size = 1.5, method = "loess"), has.band = TRUE, band = list(y = c(-2,
  2), linetype = 2, size = 0.5), is.draft = TRUE, draft = list(size = 20,
  label = "DRAFT", color = "grey50"), discrete = FALSE, ...)
}
\arguments{
\item{labels}{list of labels, like title, subtitle, xlab , ylab}

\item{axis.title}{list of axis title parameter : font size}

\item{axis.text}{list of axis title parameter : font size}

\item{ranges}{limits of x/y ranges}

\item{has.smooth}{logical if set to TRUE add smooth layer}

\item{smooth}{smooth layer parameters}

\item{has.band}{logical if TRUE add horizontal band}

\item{band}{horizontal band parameters}

\item{is.draft}{logical if TRUE add draft layer}

\item{draft}{draft layer parameters}

\item{discrete}{logical if TRUE x axis is discrete(FALSE by default)}

\item{...}{extra arguments (not used yet)}
}
\value{
An object of class \code{"pmx_gpar"}.
}
\description{
Handling pmx Graphical parameters
}
\details{
This object contains all general graphic settings. It used internally by all
pmx_plot(generic fucntion) to set the default behavior.
}
