% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_tables.R
\name{pmx_index_table}
\alias{pmx_index_table}
\title{Index table columns}
\usage{
pmx_index_table(x)
}
\arguments{
\item{x}{A list containing the tables (`x$data`) to be 
combined along with their respective names (`x$name`).}
}
\value{
A tibble of the index.
}
\description{
Index table columns
}
\keyword{internal}
