% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_is.nm.table.list}
\alias{pmx_is.nm.table.list}
\title{Test for nm_table_list class}
\usage{
pmx_is.nm.table.list(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `nm_table_list` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is a `nm_table_list` object
}
\keyword{internal}
