% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_list_nm_tables.R
\name{pmx_list_nm_tables}
\alias{pmx_list_nm_tables}
\title{List NONMEM output tables}
\usage{
pmx_list_nm_tables(nm_model = NULL)
}
\arguments{
\item{nm_model}{An nm_model object generated with \code{\link{pmx_read_nm_model}}.}
}
\description{
List NONMEM output tables file names from a \code{nm_model} object.
}
\examples{
\dontrun{
pmx_read_nm_model(file = 'run001.lst') \%>\% 
  pmx_list_nm_tables()
}

}
\seealso{
\code{\link{pmx_read_nm_model}}, \code{\link{pmx_read_nm_tables}}
}
