% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_manual_import.R
\name{pmx_list_nm_tables_manual}
\alias{pmx_list_nm_tables_manual}
\title{Creates an nm_table_list from manually defined table name patterns}
\usage{
pmx_list_nm_tables_manual(runno = NULL, file = NULL, dir = NULL, tab_list)
}
\arguments{
\item{runno}{Run number to be used to generate model file name.}

\item{file}{Model file name containing the file extension.}

\item{dir}{Location of the model files.}

\item{tab_list}{A list of table definition generated by `pmx_manual_nm_import`.}
}
\value{
A `nm_table_list`
}
\description{
Creates an nm_table_list from manually defined table name patterns
}
\keyword{internal}
