% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_tables.R
\name{pmx_merge_firstonly}
\alias{pmx_merge_firstonly}
\title{Merge firstonly table with full length tables}
\usage{
pmx_merge_firstonly(x, quiet)
}
\arguments{
\item{x}{A list containing the tables (`x$data`) to be 
merged, the firstonly flag (`x$firstonly`) and the 
indexes (`x$index`).}

\item{quiet}{Should messages be displayed to the console.}
}
\value{
A list containing `data` and `index` of the merged table.
}
\description{
Merge firstonly table with full length tables
}
\keyword{internal}
