% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{pmx_mlx}
\alias{pmx_mlx}
\title{Wrapper to pmx constructor}
\usage{
pmx_mlx(config, directory)
}
\arguments{
\item{config}{the config name}

\item{directory}{the data directory (working diectory)}
}
\value{
\code{pmxClass} object
}
\description{
Wrapper to pmx constructor
}
