% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_msg}
\alias{pmx_msg}
\title{Message function}
\usage{
pmx_msg(txt, quiet = TRUE)
}
\arguments{
\item{txt}{A string for the message.}

\item{quiet}{Should messages be displayed to the console.}
}
\value{
Silent when quiet is `TRUE` or a message is quiet is `FALSE`.
}
\description{
Message function with quiet option inspired from `ronkeizer/vpc`.
}
\keyword{internal}
