% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{pmx_nlmixr}
\alias{pmx_nlmixr}
\title{Creates pmx controller from  an nlimxr fit object}
\usage{
pmx_nlmixr(fit, dvid, conts, cats, strats, endpoint, settings, vpc = FALSE)
}
\arguments{
\item{fit}{nlmixr object}

\item{dvid}{\emph{[Optional]} \code{character} observation type parameter.}

\item{conts}{\emph{[Optional]}\code{character} vector of continuous covariates}

\item{cats}{\emph{[Optional]}\code{character} vector of categorical covariates}

\item{strats}{\emph{[Optional]}\code{character} extra stratification variables}

\item{endpoint}{\code{pmxEndpointClass} or \code{integer} or \code{charcater} defalut to NULL
of the endpoint code.   \code{\link{pmx_endpoint}}}

\item{settings}{\emph{[Optional]}\code{pmxSettingsClass} \code{\link{pmx_settings}}}

\item{vpc}{\emph{[Optional]} \code{logical} a boolean indiacting if vpc should be calculated (by default \code{TRUE})}
}
\value{
\code{pmxClass} controller object.
}
\description{
Creates pmx controller from  an nlimxr fit object
}
