% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_reader.R
\name{pmx_nm}
\alias{pmx_nm}
\title{Creates pmx controller from  NONMEM output}
\usage{
pmx_nm(runno = NULL, table_suffix = "", sim.suffix = "sim",
  cwres.suffix = "", directory = ".", quiet = TRUE,
  table_names = c("sdtab", "mutab", "patab", "catab", "cotab", "mytab",
  "extra", "xptab", "cwtab"), cwres.name = c("cwtab"),
  mod.prefix = "run", mod.suffix = ".mod", phi.suffix = ".phi",
  phi.file = NULL, nm7 = NULL, dvid, conts, cats, strats, endpoint,
  settings, vpc = FALSE, pred = "PRED", bloq)
}
\arguments{
\item{runno}{}

\item{table_suffix}{}

\item{sim.suffix}{}

\item{cwres.suffix}{}

\item{directory}{}

\item{quiet}{}

\item{table_names}{}

\item{cwres.name}{}

\item{mod.prefix}{}

\item{mod.suffix}{}

\item{phi.suffix}{}

\item{phi.file}{}

\item{nm7}{}

\item{dvid}{}

\item{conts}{}

\item{cats}{}

\item{strats}{}

\item{endpoint}{}

\item{settings}{}

\item{vpc}{}

\item{pred}{}

\item{bloq}{}
}
\description{
Creates pmx controller from  NONMEM output
}
