% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_files.R
\name{pmx_parse_nm_files}
\alias{pmx_parse_nm_files}
\title{Parse NONMEM output files}
\usage{
pmx_parse_nm_files(dat, quiet)
}
\arguments{
\item{dat}{A list containing the raw data as vector of strings (`dat$raw`)
and their respective file names (`dat$name`).}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A tibble containing the parsed `data` along with `name`, `problem`, 
`subprob`, and `method`.
}
\description{
Function parsing NONMEM output files from their 
raw input.
}
\keyword{internal}
