% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-saem-convergence.R
\name{pmx_plot_saem_convergence}
\alias{pmx_plot_saem_convergence}
\title{SAEM Convergence Plot}
\usage{
pmx_plot_saem_convergence(ctr, ...)
}
\arguments{
\item{ctr}{\code{pmxClass} controller.}

\item{...}{additional parameters (not used).}
}
\value{
A \code{ggplot} object showing SAEM convergence plot.
}
\description{
This plot displays the sequence of estimates for population parameters
computed after each iteration of the SAEM algorithm. The purpose is to check
the convergence of the algorithm. In addition, a convergence indicator gives
the estimation for -2 x log-likelihood along the iterations.
}
