% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_tables.R
\name{pmx_read_args}
\alias{pmx_read_args}
\title{Define data import arguments}
\usage{
pmx_read_args(
  x,
  quiet,
  col_types = readr::cols(.default = "d"),
  na = "NA",
  comment = "TABLE",
  skip = 1,
  ...
)
}
\arguments{
\item{x}{A list containing a the 3 first records of a 
dataset under `[[1]]`.}

\item{quiet}{Should messages be displayed to the console.}

\item{col_types}{Defines the type of each column to be passed to 
the `readr` import function.}

\item{na}{Character string defining the values to be treated as `NA`.}

\item{comment}{Character string defining the value to mark comments.}

\item{skip}{Number of rows to be skipped before reading the data.}

\item{...}{Additional arguments to be passed to the `readr` function}
}
\value{
A list of 2 levels fun (the import function) and params (a list 
of arguments to be used when calling fun).
}
\description{
Define data import arguments
}
\keyword{internal}
