% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_read_nm_tables.R
\name{pmx_read_funs}
\alias{pmx_read_funs}
\title{Define data import functions}
\usage{
pmx_read_funs(fun)
}
\arguments{
\item{fun}{Abbreviated `readr` data import function. 
Can be `csv`, `csv2` or `table`.}
}
\value{
A data import function.
}
\description{
Define data import functions
}
\keyword{internal}
