% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-all-plots.R
\name{pmx_register_plot}
\alias{pmx_register_plot}
\title{Register plot  (Jun2025, Alex: I believe it doesn't work at all)}
\usage{
pmx_register_plot(ctr, pp, pname = NULL)
}
\arguments{
\item{ctr}{\code{pmxClass} controller}

\item{pp}{\code{ggplot2} plot}

\item{pname}{\code{character} plot nme}
}
\description{
Register plot  (Jun2025, Alex: I believe it doesn't work at all)
}
