% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-report.R
\name{pmx_report_template}
\alias{pmx_report_template}
\title{Gets build-in report templates}
\usage{
pmx_report_template()
}
\value{
list of templates names
}
\description{
Gets build-in report templates
}
\examples{
pmx_report_template()
}
