% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{pmx_shrink}
\alias{pmx_shrink}
\title{Create shrinkage parameter object}
\usage{
pmx_shrink(
  fun = c("var", "sd"),
  size = 1,
  color = "green",
  vjust = 1.5,
  hjust = 0.5,
  ...
)
}
\arguments{
\item{fun}{\code{list} shrinkage function can be \code{sd} or \code{var}}

\item{size}{\code{numeric} shrinkage text size}

\item{color}{\code{character} shrinkage text color}

\item{vjust}{\code{numeric} shrinkage position vertical adjustment}

\item{hjust}{\code{numeric} shrinkage position horizontal adjustment}

\item{...}{any other parameter}
}
\value{
\code{pmxShrinkClass} object (\code{list})
}
\description{
Create shrinkage parameter object
}
