% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-themes.R
\name{pmx_theme}
\alias{pmx_theme}
\title{Define ggPMX theme}
\usage{
pmx_theme(...)
}
\arguments{
\item{...}{can contain any valid argument of \code{ggplot2} \code{\link[ggplot2]{theme}}
object.}
}
\value{
ggplot2 theme object
}
\description{
This theme is a simple wrapper gdoc theme
from ggthems package.
}
