% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_tidyr_new_interface}
\alias{pmx_tidyr_new_interface}
\title{Check tidyr version}
\usage{
pmx_tidyr_new_interface()
}
\value{
the package version of tidyr
}
\description{
Check the version of tidyr to handle the gap between v0.8.3 and v1.0.0
}
\keyword{internal}
