% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_utils_nonmem_reader.R
\name{pmx_update_extension}
\alias{pmx_update_extension}
\title{Update file extension}
\usage{
pmx_update_extension(x, ext)
}
\arguments{
\item{x}{A string or vector of strings containing the file name to be modified.}

\item{ext}{A string or vector of strings containing the name of the new extension(s).}
}
\value{
A string or vector of strings of file name(s).
}
\description{
Change the extension of a file.
}
\keyword{internal}
