% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-vpc.R
\name{pmx_vpc}
\alias{pmx_vpc}
\title{Creates vpc object}
\usage{
pmx_vpc(
  type = c("percentile", "scatter"),
  idv = "TIME",
  obs = pmx_vpc_obs(),
  pi = pmx_vpc_pi(),
  ci = pmx_vpc_ci(),
  rug = pmx_vpc_rug(),
  bin = pmx_vpc_bin(),
  labels = NULL,
  facets = NULL,
  is.legend = TRUE,
  is.footnote = TRUE,
  dname = NULL,
  ...
)
}
\arguments{
\item{type}{\code{character} can be either percentile or scatter}

\item{idv}{\code{character} individual variable}

\item{obs}{\code{pmx_vpc_obs} object observation layer \link{pmx_vpc_obs}}

\item{pi}{\code{pmx_vpc_pi} object percentile layer \link{pmx_vpc_pi}}

\item{ci}{\code{pmx_vpc_ci} object confidence interval layer \link{pmx_vpc_ci}}

\item{rug}{\code{pmx_vpc_rug} object rug layer \link{pmx_vpc_rug}}

\item{bin}{\code{pmx_vpc_bin} object \link{pmx_vpc_bin}}

\item{labels}{\code{list} define title and axis labels}

\item{facets}{is a list of parameters passed to facet_wrap in case of startification}

\item{is.legend}{\code{logical} if TRUE add legend}

\item{is.footnote}{\code{logical} if TRUE add footnote}

\item{dname}{added for compatibility with other ggPMX plots}

\item{...}{extra parameters passed to base graphical parameters}
}
\value{
list with parameters of the vpc object
}
\description{
Creates vpc object
}
\seealso{
Other vpc: 
\code{\link{pmx_plot_vpc}()},
\code{\link{pmx_vpc_bin}()},
\code{\link{pmx_vpc_ci}()},
\code{\link{pmx_vpc_obs}()},
\code{\link{pmx_vpc_pi}()},
\code{\link{pmx_vpc_rug}()}
}
\concept{vpc}
