% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-vpc.R
\name{pmx_vpc_rug}
\alias{pmx_vpc_rug}
\title{Sets vpc rug layer}
\usage{
pmx_vpc_rug(show = TRUE, color = "#000000", linewidth = 1, alpha = 0.7, size)
}
\arguments{
\item{show}{\code{logical} If TRUE show bin separators}

\item{color}{\code{character} Color of the rug. Default: "#000000".}

\item{linewidth}{\code{numeric} Thickness of the rug. Default: 1.}

\item{alpha}{\code{numeric} Transparency of the rug. Default: 0.7.}

\item{size}{\code{numeric} Depreciated thickness of the rug. Default: 1.}
}
\value{
list with options for the rug layer
}
\description{
Sets vpc rug layer
}
\details{
When the vpc confidence interval layer method is rectangles we don't show rug separators.
}
\seealso{
Other vpc: 
\code{\link{pmx_plot_vpc}()},
\code{\link{pmx_vpc}()},
\code{\link{pmx_vpc_bin}()},
\code{\link{pmx_vpc_ci}()},
\code{\link{pmx_vpc_obs}()},
\code{\link{pmx_vpc_pi}()}
}
\concept{vpc}
