% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPMX-options.R
\name{pmxOptions}
\alias{pmxOptions}
\title{This function can be used to set ggPMX options}
\usage{
pmxOptions(...)
}
\arguments{
\item{...}{Options to set, with the form \code{name = value}.}
}
\description{
\code{getPmxOption} retrieves the value of a ggPMX option.
\code{ggPMXOptions} sets the value of ggPMX options; it can also be used to
return a list of all currently-set ggPMX options.
}
\details{
There is a global option set, which is available by default.
}
\section{Options used in ggPMX}{


\itemize{
\item \strong{template_dir:} path to template directory
}
}

\examples{
\dontrun{
pmxOptions(template_dir = PATH_TO_CUSTOM_CONFIGURATION)
}
}
