% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{print.abbreviation}
\alias{print.abbreviation}
\title{S3 print abbreviation}
\usage{
\method{print}{abbreviation}(x, ...)
}
\arguments{
\item{x}{object of class configs}

\item{...}{pass additional options (not used presently)}
}
\value{
print abbreviation
}
\description{
S3 print abbreviation
}
