% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.configs}
\alias{print.configs}
\title{S3 method to print configs}
\usage{
\method{print}{configs}(x, ...)
}
\arguments{
\item{x}{object of class configs}

\item{...}{pass additonal options (not used presently)}
}
\value{
print result
}
\description{
S3 method to print configs
}
