% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{print.pmx_gpar}
\alias{print.pmx_gpar}
\title{Print pmx_gpar object}
\usage{
\method{print}{pmx_gpar}(x, ...)
}
\arguments{
\item{x}{pmx_gpar object}

\item{...}{argument passed to print ( to satisfy generic)}
}
\value{
a character description of graphical parameters
}
\description{
Print pmx_gpar object
}
