% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{print.pmxClass}
\alias{print.pmxClass}
\title{Print pmxClass object}
\usage{
\method{print}{pmxClass}(x, ...)
}
\arguments{
\item{x}{pmxClass object}

\item{...}{additinal arguments to pass to print}
}
\value{
print object to screen
}
\description{
Print pmxClass object
}
