% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.pmxConfig}
\alias{print.pmxConfig}
\title{S3 method print pmxConfig object}
\usage{
\method{print}{pmxConfig}(x, ...)
}
\arguments{
\item{x}{pmxConfig object}

\item{...}{addtional arguments to pass to print (unused currently)}
}
\value{
invisible object
}
\description{
S3 method print pmxConfig object
}
