% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_files.R
\name{raw_to_tibble}
\alias{raw_to_tibble}
\title{Convert raw strings to tibble}
\usage{
raw_to_tibble(x, sep, file)
}
\arguments{
\item{x}{A list containing the raw data as vector of strings (`x$raw`).}

\item{sep}{A separator which will be used to create columns.}

\item{file}{The name of the file to be parsed.}
}
\value{
A tibble.
}
\description{
Convert raw data strings to a tibble format.
}
\keyword{internal}
