% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-reader.R
\name{read_input}
\alias{read_input}
\title{Read Modelling input data}
\usage{
read_input(
  ipath,
  dv,
  dvid,
  cats = "",
  conts = "",
  strats = "",
  occ = "",
  endpoint = NULL,
  id = NULL,
  time = NULL
)
}
\arguments{
\item{ipath}{full path of the input file}

\item{dv}{\code{character} the name of measurable variable used in the input modelling file}

\item{dvid}{\code{character} observation type parameter}

\item{cats}{\emph{[Optional]}\code{character} vector of categorical covariates}

\item{conts}{\emph{[Optional]}\code{character} vector of continuous covariates}

\item{strats}{\emph{[Optional]}\code{character} extra stratification variables}

\item{occ}{\emph{[Optional]}\code{character} inter individual occasion variables}

\item{endpoint}{\code{integer}  null in case of a single endpoint otherwise the index of endpoints.}

\item{id}{\code{character} the name of identifier variable used in the input modelling file.}

\item{time}{\code{character} the name of time variable used in the input modelling file}
}
\value{
data.table well formatted containing modelling input data
}
\description{
Read Modelling input data
}
