% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-reader.R
\name{read_mlx_ind_est}
\alias{read_mlx_ind_est}
\title{Read MONOLIX individual parameters}
\usage{
read_mlx_ind_est(path, x, ...)
}
\arguments{
\item{path}{character path to the file}

\item{x}{dataset object}

\item{...}{extra parameter not used}
}
\value{
data.table object
}
\description{
Read MONOLIX individual parameters
}
