% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_mlx_par_est}
\alias{read_mlx_par_est}
\title{Read MONOLIX parameter estimation file}
\usage{
read_mlx_par_est(path, x)
}
\arguments{
\item{path}{character path to the file}

\item{x}{dataset object}
}
\value{
data.table object
}
\description{
Read MONOLIX parameter estimation file
}
