% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-reader.R
\name{read_mlx_pred}
\alias{read_mlx_pred}
\title{Read MONOLIX model predictions}
\usage{
read_mlx_pred(path, x, ...)
}
\arguments{
\item{path}{character path to the file}

\item{x}{dataset object}

\item{...}{extra paramter not used}
}
\value{
data.table object
}
\description{
Read MONOLIX model predictions
}
