% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-saem-convergence.R
\name{read_mlx_saem_conv}
\alias{read_mlx_saem_conv}
\title{Read MONOLIX SAEM convergence file}
\usage{
read_mlx_saem_conv(path, ...)
}
\arguments{
\item{path}{string specifying data path folder}

\item{...}{extra parameter not used}
}
\value{
data.table object
}
\description{
Read MONOLIX SAEM convergence file
}
