% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmext.R
\docType{data}
\name{read_nmext}
\alias{read_nmext}
\title{Reads .ext files generated by NONMEM}
\format{An object of class \code{help_files_with_topic} of length 1.}
\usage{
read_nmext
}
\arguments{
\item{run}{run a run number or run identifier}

\item{project}{project the NONMEM project directory}

\item{file}{file the `ext` file name}

\item{path}{path full path and file name for `ext` file}

\item{read_fun}{read_fun function to read the `ext` file}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A list with param, omega, and sigma in a format 
ready to be used.
}
\description{
Reads .ext files generated by NONMEM
}
\examples{
project <- system.file("nonmem", package = "mrgsolve") 
est <- read_nmext(1005, project = project) 
}
\keyword{datasets}
