% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residual.R
\name{residual}
\alias{residual}
\title{Create a residual object}
\usage{
residual(x, y, labels = NULL, point = NULL, ...)
}
\arguments{
\item{x}{x axis aesthetics}

\item{y}{y axis aesthetics}

\item{labels}{list that contain title,subtitle, axis labels}

\item{point}{geom point graphical parameters}

\item{...}{others graphics arguments passed to \code{\link{pmx_gpar}} internal object.}
}
\value{
a residual object
}
\description{
Create a residual object
}
\seealso{
\code{\link{plot_pmx.residual}}
}
