% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{set_abbrev}
\alias{set_abbrev}
\title{update or add a new abbreviation}
\usage{
set_abbrev(ctr, ...)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}

\item{...}{Options to set or add, with the form \code{name = value}.}
}
\description{
update or add a new abbreviation
}
\examples{
ctr <- theophylline()
ctr \%>\% set_abbrev("new_param" = "new value")
ctr \%>\% get_abbrev("new_param")
}
