% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{set_plot}
\alias{set_plot}
\title{Create a new plot  of the desired type}
\usage{
set_plot(ctr, ptype, pname, ...)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}

\item{ptype}{plot type can be:
\itemize{
 \item{"IND"}{ Individual plot type: \code{\link{individual}} }
 \item{"DIS"}{ Distribution plot type : \code{\link{distrib}}}
 \item{"RES"}{ Residual plot type :\code{\link{residual}}}
}}

\item{pname}{plot name, if missing it will be created using function aestetics}

\item{...}{other plot parameters to configure \code{"pmx_gpar"}.}
}
\value{
invisible ctr object
}
\description{
Create a new plot  of the desired type
}
