% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{shrinkage}
\alias{shrinkage}
\title{Compute shrinkage}
\usage{
shrinkage(par_est, ind_pred, sys = "mlx", fun = "sd")
}
\arguments{
\item{par_est}{data set}

\item{ind_pred}{data set}

\item{sys}{data system}

\item{fun}{can mean the sd or var}
}
\value{
data.table computing the shrinkage by effect
}
\description{
Compute shrinkage
}
