% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-individual.R
\name{split_footnote}
\alias{split_footnote}
\title{Split footnote if it exceeds maximumum width}
\usage{
split_footnote(s, n)
}
\arguments{
\item{s}{\code{character} the footnoote text}

\item{n}{\code{integer} the maximum width of a footnote}
}
\description{
Split footnote if it exceeds maximumum width
}
\examples{
split_footnote("Source: /tmp/mylongdirectory", 10)
}
