% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{[.pmx_gpar}
\alias{[.pmx_gpar}
\title{Method for subsetting "pmx_gpar" objects}
\usage{
\method{[}{pmx_gpar}(x, index, ...)
}
\arguments{
\item{x}{pmx_gpar object}

\item{index}{can be character/integer of element}

\item{...}{other parameter (not used just for generic)}
}
\value{
if exists the parameter description
}
\description{
Method for subsetting "pmx_gpar" objects
}
