% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theophylline}
\alias{theophylline}
\title{Creates pmx controller using theophylline data}
\usage{
theophylline(settings = NULL, ...)
}
\arguments{
\item{settings}{\code{pmxSettings} object}

\item{...}{other parameters of pmx_mlx like endpoint}
}
\value{
pmx controller
}
\description{
Creates pmx controller using theophylline data
}
\examples{
\dontrun{
theophylline()
}
}
