% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem_reader.R
\name{tidyr_new_interface}
\alias{tidyr_new_interface}
\title{Check tidyr version}
\usage{
tidyr_new_interface()
}
\value{
the package version of tidyr
}
\description{
Check the version of tidyr to handle the gap between v0.8.3 and v1.0.0
}
\keyword{internal}
