% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{update_plot}
\alias{update_plot}
\title{Get plot object}
\usage{
update_plot(ctr, pname, ..., pmxgpar = NULL)
}
\arguments{
\item{ctr}{\code{pmxClass} controller object}

\item{pname}{characater the plot name to update}

\item{...}{others graphical parametrs given to set the plot}

\item{pmxgpar}{a object of class pmx_gpar possibly the output of the
\code{\link{pmx_gpar}} function.}
}
\value{
ggplot object
}
\description{
Get plot object
}
