% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distribution.R
\name{wrap_formula}
\alias{wrap_formula}
\title{merge facets formula with new formula}
\usage{
wrap_formula(x, origin = "lfacet")
}
\arguments{
\item{x}{\code{formula} object}

\item{origin}{the origin formula default to ~lfacets}
}
\value{
\code{formula} object
}
\description{
merge facets formula with new formula
}
