% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxClass.R
\name{get_plot_config}
\alias{get_plot_config}
\title{Get the plot config by name}
\usage{
get_plot_config(ctr, pname)
}
\arguments{
\item{ctr}{the controller object}

\item{pname}{the plot name}
}
\value{
the config object
}
\description{
Get the plot config by name
}
\examples{
\donttest{
ctr <- theophylline()
ctr \%>\% set_plot("IND", pname = "indiv1")
ctr \%>\% get_plot_config("distr1")
}
}
\seealso{
Other pmxclass: 
\code{\link{get_cats}()},
\code{\link{get_conts}()},
\code{\link{get_covariates}()},
\code{\link{get_data}()},
\code{\link{get_occ}()},
\code{\link{get_plot}()},
\code{\link{get_strats}()},
\code{\link{plot_names}()},
\code{\link{plots}()},
\code{\link{pmx_update}()},
\code{\link{set_data}()},
\code{\link{set_plot}()}
}
\concept{pmxclass}
