% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPMX-package.R
\docType{package}
\name{ggPMX}
\alias{ggPMX-package}
\alias{ggPMX}
\title{ggPMX: A ggplot2 toolbox for Nonlinear Mixed-Effect Model graphical}
\description{
This package aims to generate diagnostic plots in a standard way. The tool reads data
from many sources (MONOLIX,NONMEM,others) and generates standard grahics
that can be easily integrated in a single report.
}
\details{
\itemize{
\item Get data from different system and create a data source
\item Plot many plots using the generic plot method \code{\link{plot_pmx}}.
}

For support, feedback or bug reports, please reach out to
\email{ggPMX_ORG@dl.mgd.novartis.com}.
}
\section{Version History}{


\describe{

\item{Jan 11 2017, 0.0.0}{Init ggPMX from Novartis rtemplate.}
\item{Feb 06 2017, 0.3.0}{Import version 0.3.0 of package.}

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggPMXdevelopment/ggPMX}
  \item Report bugs at \url{https://github.com/ggPMXdevelopment/ggPMX/issues}
}

}
\author{
\strong{Maintainer}: Aleksandr Pogodaev \email{alex.pogodaev@novartis.com} (\href{https://orcid.org/0000-0002-1371-207X}{ORCID})

Authors:
\itemize{
  \item Amine Gassem \email{contact@ag-study.com}
  \item Bruno Bieth \email{biethb@hotmail.com}
  \item Irina Baltcheva \email{irina.baltcheva@gmail.com}
  \item Thomas Dumortier \email{ thomas.dumortier@novartis.com}
  \item Christian Bartels \email{christian.bartels@novartis.com}
  \item Souvik Bhattacharya \email{souvik.bhattacharya@novartis.com}
  \item Inga Ludwig \email{inga.ludwig@novartis.com}
  \item Ines Paule \email{ines.paule@novartis.com}
  \item Didier Renard \email{didier.renard@novartis.com}
  \item Matthew Fidler \email{matt.fidler@novartis.com} (\href{https://orcid.org/0000-0001-8538-6691}{ORCID})
  \item Seid Hamzic \email{se.hamzic@gmail.com}
  \item Danielle Navarro \email{djnavarro@protonmail.com}
}

Other contributors:
\itemize{
  \item Benjamin Guiastrennec \email{guiastrennec@gmail.com} [contributor]
  \item Kyle T Baron \email{kyleb@metrumrg.com} (\href{https://orcid.org/0000-0001-7252-5656}{ORCID}) [contributor]
  \item Qing Xi Ooi \email{ooi.qingxi@gmail.com} [contributor]
  \item Ibtissem Rebai \email{ibtissem.rebai@certara.com} [contributor]
  \item Mahmoud Ali \email{mahmoud.ali@certara.com} [contributor]
  \item Novartis Pharma AG [copyright holder]
}

}
\keyword{internal}
