% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual.R
\name{individual}
\alias{individual}
\title{This function can be used to obtain individual prediction and compare with observed data and population prediction
for each individual separately}
\usage{
individual(labels, facets = NULL, dname = NULL, ipred_line = NULL,
  pred_line = NULL, point = NULL, bloq = NULL, is.legend,
  use.finegrid, ...)
}
\arguments{
\item{labels}{plot texts. labels, axis,}

\item{facets}{list facets settings nrow/ncol}

\item{dname}{name of dataset to be used}

\item{ipred_line}{\code{list} some pred line geom properties aesthetics}

\item{pred_line}{\code{list} some ipred line geom properties aesthetics}

\item{point}{\code{list} some point geom properties aesthetics}

\item{bloq}{\code{pmxBLOQ} object created by \code{\link{pmx_bloq}}}

\item{is.legend}{\code{logical} if TRUE add a legend}

\item{use.finegrid}{\code{logical} if FALSE use predictions data set}

\item{...}{others graphics arguments passed to \code{\link{pmx_gpar}} internal object.}
}
\value{
individual fit object
}
\description{
This function can be used to obtain individual prediction and compare with observed data and population prediction
for each individual separately
}
\seealso{
\code{\link{plot_pmx.individual}}

Other plot_pmx: \code{\link{distrib}},
  \code{\link{eta_cov}}, \code{\link{eta_pairs}},
  \code{\link{plot_pmx.distrib}},
  \code{\link{plot_pmx.eta_cov}},
  \code{\link{plot_pmx.eta_pairs}},
  \code{\link{plot_pmx.individual}},
  \code{\link{plot_pmx.pmx_dens}},
  \code{\link{plot_pmx.pmx_gpar}},
  \code{\link{plot_pmx.pmx_qq}},
  \code{\link{plot_pmx.residual}}, \code{\link{plot_pmx}}
}
\concept{plot_pmx}
