% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Obtain the data source config}
\usage{
load_config(x, sys = c("mlx", "nm", "mlx18"))
}
\arguments{
\item{x}{the config name.}

\item{sys}{can be mlx,nm,...}
}
\value{
a list :data configuration object
}
\description{
Obtain the data source config
}
