% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-reader.R
\name{load_data_set}
\alias{load_data_set}
\title{Load data set}
\usage{
load_data_set(x, path, sys, ...)
}
\arguments{
\item{x}{data set config}

\item{path}{character path to the directory}

\item{sys}{\code{character} mlx or nm}

\item{...}{extra parameter passed to special readers}
}
\value{
data.table
}
\description{
Load data set
}
