% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-reader.R
\name{load_source}
\alias{load_source}
\title{Load all/or some source data set}
\usage{
load_source(sys, path, dconf, ...)
}
\arguments{
\item{sys}{type cane mlx/nom}

\item{path}{\code{character} directory path containing all sources.}

\item{dconf}{configuration object}

\item{...}{any extra parameters for readers}
}
\value{
list of data.table
}
\description{
Load all/or some source data set
}
