% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-individual.R
\name{plot_pmx.individual}
\alias{plot_pmx.individual}
\title{This function can be used to plot individual prediction and compare with observed data and population prediction
for each individual separately}
\usage{
\method{plot_pmx}{individual}(x, dx, ...)
}
\arguments{
\item{x}{individual object}

\item{dx}{data set}

\item{...}{not used for the moment}
}
\value{
a list of ggplot2
}
\description{
This function can be used to plot individual prediction and compare with observed data and population prediction
for each individual separately
}
\seealso{
Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()},
\code{\link{plot_pmx_gpar_real}()},
\code{\link{plot_pmx}()}
}
\concept{plot_pmx}
