% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-base.R
\name{plot_pmx.pmx_gpar}
\alias{plot_pmx.pmx_gpar}
\title{The ggPMX base plot function}
\usage{
\method{plot_pmx}{pmx_gpar}(x, dx, ...)
}
\arguments{
\item{x}{object of pmx_gpar type}

\item{dx}{plot}

\item{...}{ignored parameters}
}
\value{
ggplot2 object
}
\description{
This function should be called internally by other plots to set
general settings like , smoothing, add band, labelling, theming,...
}
\seealso{
Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_qq}()},
\code{\link{plot_pmx.residual}()},
\code{\link{plot_pmx}()}
}
\concept{plot_pmx}
