% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_pmx}
\alias{plot_pmx}
\title{Generic plot method for pmx objects}
\usage{
plot_pmx(x, dx, ...)
}
\arguments{
\item{x}{object to plot}

\item{dx}{data.table , plot source data}

\item{...}{extra argument (not used)}
}
\description{
Generic plot method for pmx objects
}
\seealso{
\code{\link{pmx_gpar}}.

Other plot_pmx: \code{\link{plot_pmx.distrib}},
  \code{\link{plot_pmx.individual}},
  \code{\link{plot_pmx.residual}}
}
