% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residual.R
\name{plot_pmx.residual}
\alias{plot_pmx.residual}
\title{This function plots residual for each observed value by finding the difference between observed and predicted points. It also fits a distribution to the residual value.}
\usage{
\method{plot_pmx}{residual}(x, dx, ...)
}
\arguments{
\item{x}{residual object}

\item{dx}{data set}

\item{...}{not used for the moment}
}
\value{
ggplot2 object
}
\description{
This function plots residual for each observed value by finding the difference between observed and predicted points. It also fits a distribution to the residual value.
}
\seealso{
\code{\link{residual}}

Other plot_pmx: 
\code{\link{distrib}()},
\code{\link{eta_cov}()},
\code{\link{eta_pairs}()},
\code{\link{individual}()},
\code{\link{plot_pmx}()},
\code{\link{plot_pmx.distrib}()},
\code{\link{plot_pmx.eta_cov}()},
\code{\link{plot_pmx.eta_pairs}()},
\code{\link{plot_pmx.individual}()},
\code{\link{plot_pmx.pmx_dens}()},
\code{\link{plot_pmx.pmx_gpar}()},
\code{\link{plot_pmx.pmx_param_history}()},
\code{\link{plot_pmx.pmx_qq}()}
}
\concept{plot_pmx}
