% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx-plots-vpc.R
\name{pmx_vpc_bin}
\alias{pmx_vpc_bin}
\title{Creates vpc bins}
\usage{
pmx_vpc_bin(style, within_strat = TRUE, seed = 42, ...)
}
\arguments{
\item{style}{\code{character} style    chosen on of the:\cr
"fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust" or "jenks"}

\item{within_strat}{\code{logical} if TRUE compute the bining for each strat level. \cr
By default t is false and bining are equal for all stratifications levels.}

\item{seed}{\code{integer} used in \code{set.seed} call to ensure \cr
reproducibility if style is "kmeans". Set to NULL if this \cr
is not desired.}

\item{...}{other classInt::classIntervals parameters excpet \code{style} and \code{n}}
}
\value{
list with options for `pmx_vpc_bin` object
}
\description{
Creates vpc bins
}
\details{
This is a wrapper to the bin based VPC
}
\seealso{
Other vpc: 
\code{\link{pmx_plot_vpc}()},
\code{\link{pmx_vpc}()},
\code{\link{pmx_vpc_ci}()},
\code{\link{pmx_vpc_obs}()},
\code{\link{pmx_vpc_pi}()},
\code{\link{pmx_vpc_rug}()}
}
\concept{vpc}
