% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.configs}
\alias{print.configs}
\title{This function can be used to print configuration of the defined object using S3 method.}
\usage{
\method{print}{configs}(x, ...)
}
\arguments{
\item{x}{object of class configs}

\item{...}{pass additional options (not used presently)}
}
\value{
print result
}
\description{
This function can be used to print configuration of the defined object using S3 method.
}
